refactor: PushService를 빈으로 등록
