+++ src/main/java/com/wellmeet/notification/webpush/WebPushService.java
@@ -7,8 +7,8 @@ import com.wellmeet.notification.webpush.dto.SubscribeRequest;
 import com.wellmeet.notification.webpush.dto.SubscribeResponse;
 import com.wellmeet.notification.webpush.dto.TestPushRequest;
 import com.wellmeet.notification.webpush.dto.UnsubscribeRequest;
-import com.wellmeet.notification.webpush.infrastructure.WebPushSender;
 import com.wellmeet.notification.webpush.repository.PushSubscriptionRepository;
+import com.wellmeet.notification.webpush.sender.WebPushSender;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
