+++ src/test/java/com/wellmeet/BaseControllerTest.java
@@ -1,5 +1,6 @@
 package com.wellmeet;
 
+import com.wellmeet.config.WebPushTestConfig;
 import com.wellmeet.notification.webpush.repository.PushSubscriptionRepository;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
@@ -11,11 +12,13 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
 @ExtendWith(DataBaseCleaner.class)
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@Import(WebPushTestConfig.class)
 public abstract class BaseControllerTest {
 
     @Autowired
